validaciones de permisos e importando a app.js

+++ src/controllers/product.controller.js
@@ -40,11 +40,11 @@ class ProductController {
 
   // ADD PRODUCT
   addProduct = async (req, res, next) => {
-    const { title, description, code, price, status, stock, category } = req.body;
+    const { title, description, code, price, status, stock, category, owner } = req.body;
     if (!title || !description || !code || !price || !status || !stock || !category) {
       const error = CustomError.createError({
-        name: 'InvalidData', 
-        cause: sendErrorInfo(req.body), 
+        name: 'InvalidData',
+        cause: sendErrorInfo(req.body),
         message: 'Please enter the correct data',
         code: enumErrors.INCOMPLETE_DATA
       });
@@ -53,15 +53,15 @@ class ProductController {
     else if (typeof (title) !== 'string' || typeof (description) !== 'string' ||
       typeof (code) !== 'string' || typeof (price) !== 'number' || typeof (status) !== 'boolean' ||
       typeof (stock) !== 'number' || typeof (category) !== 'string') {
-        const error = CustomError.createError({
-          name: 'Invalid types of data', 
-          cause: sendErrorInfo(req.body), 
-          message: 'Please enter the correct data',
-          code: enumErrors.ERROR_DATA_TYPES
-        });
-        return next(error);
+      const error = CustomError.createError({
+        name: 'Invalid types of data',
+        cause: sendErrorInfo(req.body),
+        message: 'Please enter the correct data',
+        code: enumErrors.ERROR_DATA_TYPES
+      });
+      return next(error);
     }
-    let statusRes = await productRepository.addProduct({ title, description, code, price, status, stock, category });
+    let statusRes = await productRepository.addProduct({ title, description, code, price, status, stock, category, owner });
     if (statusRes.code = enumErrors.CODE_EXISTS) return sendError(res, 400, 'Product already exists');
     //socketServer.emit('addProduct', { title, description, code, price, status, stock, category })    
     sendPayload(res, 200, statusRes);
@@ -71,6 +71,9 @@ class ProductController {
   updateProduct = async (req, res) => {
     const { pid } = req.params;
     const { id, ...rest } = req.body;
+    const { role, email } = req.session.user;
+    let productFind = await productRepository.findProduct(pid);
+    if (role === 'premium' && email !== productFind.owner) return sendError(res, 403, 'You cannot delete this product');
     let resUpdateProduct = await productRepository.updateProduct(pid, rest);
     if (resUpdateProduct.matchedCount === 0 || resUpdateProduct === 'CastError') return sendError(res, 400, 'Product not found');
     sendPayload(res, 200, 'Product modified successfully');
@@ -79,8 +82,10 @@ class ProductController {
   // DELETE PRODUCT
   deleteProduct = async (req, res) => {
     const { pid } = req.params;
+    const { role, email } = req.session.user;
     let productFind = await productRepository.findProduct(pid);
     if (!productFind || productFind === 'CastError') return sendError(res, 400, 'Product not found');
+    if (role === 'premium' && email !== productFind.owner) return sendError(res, 403, 'You cannot delete this product');
     await productsInstance.deleteProduct(pid);
     // socketServer.emit('deleteProduct', productFind.code)
     sendPayload(res, 200, 'Product deleted successfully');
